# contracts/openapi/superartillery.yaml
openapi: 3.1.0
info:
  title: SuperArtillery API
  version: 1.1.0
  description: Contract-first API specification for REST endpoints and server-to-client game messages.
servers:
  - url: http://localhost:3000
tags:
  - name: Health
  - name: Games
  - name: Invitations
  - name: Gameplay
paths:
  /api/v1/health:
    get:
      tags: [Health]
      summary: Enhanced health check for the server
      description: Returns server health, game count, and capacity information
      responses:
        '200':
          description: Server is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /api/v1/games:
    post:
      tags: [Games]
      summary: Create a new private game
      description: Create a new private two-player game and return an invitation code and link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGameRequest'
      responses:
        '201':
          description: Game created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateGameResponse'
        '400':
          description: Invalid player name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: Server at maximum capacity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/invitations/accept:
    post:
      tags: [Invitations]
      summary: Accept a game invitation
      description: Accept an invitation via token or short code and join the game
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInvitationRequest'
      responses:
        '200':
          description: Invitation accepted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptInvitationResponse'
        '400':
          description: Invalid invitation or player name
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '410':
          description: Invitation expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/games/{gameId}/status:
    get:
      tags: [Games]
      summary: Get game status
      description: Poll game status (requires valid session token)
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
        - name: sessionToken
          in: query
          required: true
          schema:
            type: string
          description: Player session token for authentication
      responses:
        '200':
          description: Game status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GameStatusResponse'
        '401':
          description: Invalid session token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game not found or expired
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/games/{gameId}/rematch:
    post:
      tags: [Games]
      summary: Request another round
      description: Mark the authenticated player as ready for another round after game over
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
        - name: sessionToken
          in: query
          required: true
          schema:
            type: string
          description: Player session token for authentication
      responses:
        '200':
          description: Rematch request accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RematchResponse'
        '400':
          description: Rematch is not available for the current game state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid session token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/fire:
    post:
      tags: [Gameplay]
      summary: Fire a shot (session token authenticated)
      description: Fire a projectile. Requires gameId and sessionToken. Server derives player identity from token.
      parameters:
        - name: sessionToken
          in: query
          required: true
          schema:
            type: string
          description: Player session token for authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FireRequest'
      responses:
        '200':
          description: Fire action accepted (server broadcasts shot + turn_change/game_over via WebSocket)
        '400':
          description: Validation or game-state error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid session token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Game not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    PlayerId:
      type: integer
      enum: [0, 1]

    Position:
      type: object
      additionalProperties: false
      required: [x, y]
      properties:
        x:
          type: number
        y:
          type: number

    Castle:
      type: object
      additionalProperties: false
      required: [playerId, left_x, base_y]
      properties:
        playerId:
          $ref: '#/components/schemas/PlayerId'
        left_x:
          type: number
        base_y:
          type: number

    Battlefield:
      type: object
      additionalProperties: false
      required: [canvasWidth, canvasHeight, gravity, wind, groundY, castleWidth, castleHeight, castles, terrain]
      properties:
        canvasWidth:
          type: number
        canvasHeight:
          type: number
        gravity:
          type: number
        wind:
          type: number
          minimum: -50
          maximum: 50
          description: Signed horizontal wind acceleration; negative blows left and positive blows right
        groundY:
          type: number
        castleWidth:
          type: number
        castleHeight:
          type: number
        castles:
          type: array
          minItems: 2
          maxItems: 2
          items:
            $ref: '#/components/schemas/Castle'
        terrain:
          type: object
          additionalProperties: false
          required: [version, seed, sampleWidth, minY, maxY, leftY, rightY, hillCenter, hillWidth, hillHeight]
          properties:
            version:
              type: integer
              minimum: 1
            seed:
              type: integer
              minimum: 0
            sampleWidth:
              type: number
              exclusiveMinimum: 0
            minY:
              type: number
            maxY:
              type: number
            leftY:
              type: number
            rightY:
              type: number
            hillCenter:
              type: number
            hillWidth:
              type: number
              exclusiveMinimum: 0
            hillHeight:
              type: number
              description: Signed cosine terrain variation; positive values create a crest, negative values create a depression, and zero produces a slope between the side elevations

    # Private Games Requests/Responses

    CreateGameRequest:
      type: object
      additionalProperties: false
      required: [playerName]
      properties:
        playerName:
          type: string
          minLength: 1
          maxLength: 15
          description: Display name for the initiating player (15 chars max, must start with alphanumeric)
        clientUrl:
          type: string
          format: uri
          description: Current client URL used to preserve the deployed application path in the invite link

    CreateGameResponse:
      type: object
      additionalProperties: false
      required: [gameId, playerToken, inviteUrl, inviteCode]
      properties:
        gameId:
          type: string
          description: Opaque game ID
        playerToken:
          type: string
          description: Session token for this player (high entropy, one-time)
        inviteUrl:
          type: string
          format: uri
          description: Full invitation link with the invite code and server URL (shareable)
        inviteCode:
          type: string
          minLength: 4
          maxLength: 4
          description: Short 4-character alphanumeric code (easy to type)

    AcceptInvitationRequest:
      type: object
      additionalProperties: false
      required: [playerName]
      properties:
        inviteCode:
          type: string
          minLength: 4
          maxLength: 4
          description: Short invite code from the invitation link or displayed code
        playerName:
          type: string
          minLength: 1
          maxLength: 15
          description: Display name for the invited player

    AcceptInvitationResponse:
      type: object
      additionalProperties: false
      required: [gameId, playerToken]
      properties:
        gameId:
          type: string
          description: Opaque game ID
        playerToken:
          type: string
          description: Session token for this player (high entropy, one-time)

    GameStatusResponse:
      type: object
      additionalProperties: false
      required: [status, playersConnected, requiredPlayers, rematchReady, rematchPlayersReady]
      properties:
        status:
          type: string
          enum: [pending, active, finished, expired]
          description: Current game status
        playersConnected:
          type: integer
          minimum: 0
          maximum: 2
          description: Number of players currently connected via WebSocket
        requiredPlayers:
          type: integer
          const: 2
          description: Required number of players for the game to start
        rematchReady:
          type: boolean
          description: Whether the authenticated player has requested another round
        rematchPlayersReady:
          type: integer
          minimum: 0
          maximum: 2
          description: Number of players who have requested another round

    RematchResponse:
      type: object
      additionalProperties: false
      required: [ready, playersReady, requiredPlayers, roundStarted]
      properties:
        ready:
          type: boolean
          description: Whether the authenticated player is ready for another round
        playersReady:
          type: integer
          minimum: 0
          maximum: 2
        requiredPlayers:
          type: integer
          const: 2
        roundStarted:
          type: boolean
          description: Whether both players were ready and a new round started

    FireRequest:
      type: object
      additionalProperties: false
      required: [gameId, angle, velocity]
      properties:
        gameId:
          type: string
          description: Opaque game ID
        angle:
          type: number
          minimum: 0
          maximum: 360
          description: Projectile angle in degrees
        velocity:
          type: number
          exclusiveMinimum: 0
          description: Projectile velocity (must be positive)

    ErrorResponse:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Optional additional error context

    HealthResponse:
      type: object
      additionalProperties: false
      required: [status, timestamp, uptime, gameCount, invitationCount, maxGamesReached, version, contractVersion]
      properties:
        status:
          type: string
          enum: [ok, degraded]
          description: Server status (degraded if at max capacity)
        timestamp:
          type: string
          format: date-time
        uptime:
          type: string
          pattern: '^\\d+\\.\\d{2}:\\d{2}:\\d{2}\\.\\d{3}$'
          description: Server uptime formatted as D.HH:mm:ss.ggg
        gameCount:
          type: integer
          minimum: 0
          description: Number of active games
        invitationCount:
          type: integer
          minimum: 0
          description: Number of pending invitations
        maxGamesReached:
          type: boolean
          description: True if server is at maximum capacity
        version:
          type: string
        contractVersion:
          type: string
          description: Client/server protocol contract version

    # WebSocket Message Types

    GameStartMessage:
      type: object
      additionalProperties: false
      required: [type, gameId, opponentName, battlefield, round]
      properties:
        type:
          type: string
          const: game_start
        gameId:
          type: string
          description: Opaque game ID
        opponentName:
          type: string
          description: Display name of the opponent
        battlefield:
          $ref: '#/components/schemas/Battlefield'
        round:
          type: integer
          minimum: 1

    ShotMessage:
      type: object
      additionalProperties: false
      required: [type, playerId, angle, velocity]
      properties:
        type:
          type: string
          const: shot
        playerId:
          $ref: '#/components/schemas/PlayerId'
        angle:
          type: number
        velocity:
          type: number

    TurnChangeMessage:
      type: object
      additionalProperties: false
      required: [type, playerId_turn]
      properties:
        type:
          type: string
          const: turn_change
        playerId_turn:
          $ref: '#/components/schemas/PlayerId'

    GameOverMessage:
      type: object
      additionalProperties: false
      required: [type, playerId_winner]
      properties:
        type:
          type: string
          const: game_over
        playerId_winner:
          $ref: '#/components/schemas/PlayerId'

    RematchStatusMessage:
      type: object
      additionalProperties: false
      required: [type, playersReady, requiredPlayers]
      properties:
        type:
          type: string
          const: rematch_status
        playersReady:
          type: integer
          minimum: 0
          maximum: 2
        requiredPlayers:
          type: integer
          const: 2

    WebSocketErrorMessage:
      type: object
      additionalProperties: false
      required: [type, code, message]
      properties:
        type:
          type: string
          const: error
        code:
          type: string
          description: Machine-readable WebSocket error code
        message:
          type: string
          description: Human-readable WebSocket error message
        details:
          type: object
          additionalProperties: true
          description: Optional error context

    GameMessage:
      oneOf:
        - $ref: '#/components/schemas/GameStartMessage'
        - $ref: '#/components/schemas/ShotMessage'
        - $ref: '#/components/schemas/TurnChangeMessage'
        - $ref: '#/components/schemas/GameOverMessage'
        - $ref: '#/components/schemas/RematchStatusMessage'
        - $ref: '#/components/schemas/WebSocketErrorMessage'
      discriminator:
        propertyName: type
        mapping:
          game_start: '#/components/schemas/GameStartMessage'
          shot: '#/components/schemas/ShotMessage'
          turn_change: '#/components/schemas/TurnChangeMessage'
          game_over: '#/components/schemas/GameOverMessage'
          rematch_status: '#/components/schemas/RematchStatusMessage'
          error: '#/components/schemas/WebSocketErrorMessage'